# Faster vote handling