--- a/zilliqa/src/consensus.rs
+++ b/zilliqa/src/consensus.rs
@@ -1109,7 +1109,9 @@ impl Consensus {
         }
 
         // Either way assemble early proposal now if it doesnt already exist
-        self.early_proposal_assemble_at(None)?;
+        if self.transaction_pool.has_txn_ready() {
+            self.early_proposal_assemble_at(None)?;
+        }
 
         Ok(None)
     }