--- a/zilliqa/src/consensus.rs
+++ b/zilliqa/src/consensus.rs
@@ -412,8 +412,8 @@ impl Consensus {
                 ..block.header
             };
             let root_hash = self.state.root_hash()?;
-            let stakers = self.get_stakers_mut(root_hash, next_block_header)?; // self.state.get_stakers(next_block_header)?;
-                                                                               // If we're in the genesis committee, vote again.
+            let stakers = self.get_stakers_mut(root_hash, next_block_header)?;
+            // If we're in the genesis committee, vote again.
             if stakers.iter().any(|v| *v == self.public_key()) {
                 info!("timeout in view: {:?}, we will vote for block rather than incrementing view, block hash: {}", view, block.hash());
                 let leader = self.leader_at_block(&block, view).unwrap();
@@ -691,7 +691,7 @@ impl Consensus {
             }
 
             let root_hash = self.state.root_hash()?;
-            let stakers = self.get_stakers_mut(root_hash, block.header)?; // self.state.get_stakers(block.header)?;
+            let stakers = self.get_stakers_mut(root_hash, block.header)?;
 
             // It is possible to source Proposals from own storage during sync, which alters the source of the Proposal.
             // Only allow from == self, for fast-forwarding, in normal case but not during sync
@@ -739,7 +739,7 @@ impl Consensus {
                 ..block.header
             };
             let root_hash = self.state.root_hash()?;
-            let stakers = self.get_stakers_mut(root_hash, next_block_header)?; //  self.state.get_stakers(next_block_header)?;
+            let stakers = self.get_stakers_mut(root_hash, next_block_header)?;
 
             if !stakers.iter().any(|v| *v == self.public_key()) {
                 debug!(
@@ -1030,7 +1030,6 @@ impl Consensus {
     }
 
     pub fn vote(&mut self, vote: Vote) -> Result<Option<(Block, Vec<VerifiedTransaction>)>> {
-        let now = std::time::Instant::now();
         let block_hash = vote.block_hash;
         let block_view = vote.view;
         let current_view = self.get_view()?;
@@ -1047,8 +1046,6 @@ impl Consensus {
         // words, a malicious node which is not part of the consensus committee may send us a vote and this check will
         // still pass. We later validate that the owner of `vote.public_key` is a valid voter.
         vote.verify()?;
-        debug!("VOTE_VERIFY {:?}", now.elapsed());
-        let now = std::time::Instant::now();
 
         // Retrieve the actual block this vote is for.
         let Some(block) = self.get_block(&block_hash)? else {
@@ -1062,7 +1059,6 @@ impl Consensus {
                 .push(vote);
             return Ok(None);
         };
-        debug!("VOTE_SUPERMAJORITY_1 {:?}", now.elapsed());
 
         // if we are not the leader of the round in which the vote counts
         // The vote is in the happy path (?) - so the view is block view + 1
@@ -1074,7 +1070,6 @@ impl Consensus {
             );
             return Ok(None);
         }
-        debug!("VOTE_SUPERMAJORITY_2 {:?}", now.elapsed());
 
         let executed_block = BlockHeader {
             number: block.header.number + 1,
@@ -1082,7 +1077,6 @@ impl Consensus {
         };
 
         let committee = self.get_stakers_mut(block.state_root_hash(), executed_block)?;
-        debug!("VOTE_SUPERMAJORITY_3 {:?}", now.elapsed());
 
         // verify the sender's signature on block_hash
         let Some((index, _)) = committee
@@ -1093,7 +1087,6 @@ impl Consensus {
             warn!("Skipping vote outside of committee");
             return Ok(None);
         };
-        debug!("VOTE_SUPERMAJORITY_4 {:?}", now.elapsed());
 
         let (mut signatures, mut cosigned, mut cosigned_weight, mut supermajority_reached) =
             self.votes.get(&block_hash).cloned().unwrap_or_else(|| {
@@ -1105,7 +1098,6 @@ impl Consensus {
                 )
             });
 
-        debug!("VOTE_SUPERMAJORITY_5 {:?}", now.elapsed());
         if supermajority_reached {
             trace!(
                 "(vote) supermajority already reached in this round {}",
@@ -1113,7 +1105,6 @@ impl Consensus {
             );
             return Ok(None);
         }
-        let now = std::time::Instant::now();
 
         // if the vote is new, store it
         if !cosigned[index] {
@@ -1124,13 +1115,11 @@ impl Consensus {
             let Some(weight) = self.state.get_stake(vote.public_key, executed_block)? else {
                 return Err(anyhow!("vote from validator without stake"));
             };
-            debug!("VOTE_PROPOSE_1 {:?}", now.elapsed());
 
             cosigned_weight += weight.get();
 
             let total_weight = self.total_weight(executed_block);
             supermajority_reached = cosigned_weight * 3 > total_weight * 2;
-            debug!("VOTE_PROPOSE_2 {:?}", now.elapsed());
 
             trace!(
                 cosigned_weight,
@@ -1149,7 +1138,6 @@ impl Consensus {
                     supermajority_reached,
                 ),
             );
-            debug!("VOTE_PROPOSE_3 {:?}", now.elapsed());
 
             // if we are already in the round in which the vote counts and have reached supermajority
             if supermajority_reached {
@@ -1168,15 +1156,12 @@ impl Consensus {
                 block_hash,
                 (signatures, cosigned, cosigned_weight, supermajority_reached),
             );
-            debug!("VOTE_PROPOSE_4 {:?}", now.elapsed());
         }
 
         // Either way assemble early proposal now if it doesnt already exist
         if self.transaction_pool.has_txn_ready() {
             self.early_proposal_assemble_at(None)?;
         }
-
-        debug!("VOTE_PROPOSE_N {:?}", now.elapsed());
 
         Ok(None)
     }