Analysis of the vote handling process on local, finds that in the case of a 4-validator network, it takes about 25ms to handle a vote. Of this, ~13ms is spent in `total_weight()` and ~7ms is spent in `get_stakers()` calls to the EVM. This represents about 80% of the overall vote handling time. Increasing the number of validators also increases the vote handling time (as observed earlier).

This PR introduces some changes:
1. Rewrite of `total_weight()` from O(n) to O(1) time, which reduced the handling time from 13ms to ~3ms by calling the `getTotalStake()` deposit contract; then adding a caching mechanism reduced this further to a negligible cost after the first call.
2. Adding caching to `get_stakers()`, which reduces the call cost to negligible, after the first call.

This brings into the question of whether it might make sense to build a generalised caching mechanism for all EVM calls. But that might be overkill for now, and is something to think about if the need arises.

I think it's fine to merge these changes first, since they definitely provide some measure of improvement to the vote handling process. Hopefully, we can test and measure the real impact of these changes once the present devops issues have been resolved.