@shawn-zil 2024-11-12
> @shawn-zil Do you have a benchmark or easily reproducible way of testing the vote processing speed? I'm keen on exploring options to do the caching at a slightly lower level which avoids adding even more state and complexity into `Consensus`.
> 
> I'm testing a state cache in devnet this morning - https://github.com/Zilliqa/zq2/commit/61e3ae91ed75a67917cfb06c7e0466000e565939
> 
> Can you test the same change with your specific vote-processing issue?

@JamesHinshelwood 
I'm not doing anything fancy, just using timestamp.elapsed and printing it out at different points of the function to take readings.

Sorry.