@JamesHinshelwood 2024-11-12
Once I've deployed my change to devnet, would you mind checking the performance there too to see if it helps with your issue?