@shawn-zil 2024-11-26
Also impacts #1525